requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
chromadb
pymupdf
pytesseract
Pillow
requests
tqdm
sentence-transformers
numpy
tenacity
streamlit
